Show server/client "has responded" flags.
Instead of trying to show which headers were added at what stages, let’s just actually relay a message from the server or the cache (via logfiles) down to the client when they respond to something.

This way, we’re not doing any voodoo in the client — when the server says it responded, it responded — and when the cache said it responded, it responded. Easy as pie.

Since the client is already just rendering the headers of the XMLHttpRequest that’s being made, we can safely stop parsing the headers on the server of the cache response, and just tunnel them through to the client un-modified. That’s coming up next!