Don’t show “force reload” when in a textarea.
When we’re editing the response headers, sometimes I use ALT+Arrows to select words at a time. This was a little distracting, because the “Force ” was show/hiding in the background.

This commit prevents the ALT+Click to force request when the textarea is open — no big deal, since we can’t actually click that when the dialog is open anyways.